norm check
